fixes reward probability gathering indices

+++ src/homa/rl/diayn/Discriminator.py
@@ -35,7 +35,7 @@ class Discriminator(MovesNetworkToDevice):
         entropy = numpy.log(1.0 / self.num_skills)
         if skill_index.dim() == 1:
             skill_index = skill_index.unsqueeze(-1)
-        reward = probabilities.gather(1, skill_index) - entropy
+        reward = probabilities.gather(1, skill_index.long()) - entropy
         return reward.squeeze(-1)
 
     def train(self, states: torch.Tensor, skills_indices: torch.Tensor):
